requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PyYAML==6.0
pyzabbix==1.0.0
requests==2.27.1
semantic-version==2.9.0
urllib3==1.26.8
